requirements.txt
gradio
chatterbox-tts
numpy